requirements.txt
appdirs==1.4.3
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
black==18.9b0
Click==7.0
coverage==4.5.2
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
flake8==3.6.0
gunicorn==19.9.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
mccabe==0.6.1
model-mommy==1.6.0
more-itertools==5.0.0
numpy==1.16.0
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.1
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
pytest==4.1.1
pytest-cov==2.6.1
pytest-django==3.4.5
python-dateutil==2.7.5
python-dotenv==0.10.1
python-json-logger==0.1.10
pytz==2018.9
six==1.12.0
toml==0.10.0
traitlets==4.3.2
wcwidth==0.1.7
